refactor: creating CurrencyInput to be specific for home page and improve UX
